First round of fixes after static analysis (Coverity)

+++ src/utils.c
@@ -148,7 +148,7 @@ uint64_t imquic_read_pfxint(uint8_t n, uint8_t *bytes, size_t blen, uint8_t *len
 			return 0;
 		}
 		b = bytes[i];
-		number += (b & 0x7f) * (1 << m);
+		number += (b & 0x7f) * ((uint64_t)1 << m);
 		m += 7;
 	};
 	if(length)
